test/unit/test_scraper.py
"""
Unit tests for the Scraper module.
"""

import pytest
import tempfile
import os
from pathlib import Path
from scraper import ScraperModule, ScrapeResult


def test_scraper_initialization():
    """Test basic scraper module initialization"""
    with tempfile.TemporaryDirectory() as temp_dir:
        output_dir = Path(temp_dir) / "scrapes"
        scraper = ScraperModule(
            delay_between_requests=0.1,
            timeout=10,
            output_dir=str(output_dir)
        )

        assert scraper.delay_between_requests == 0.1
        assert scraper.timeout == 10
        assert output_dir.exists()


def test_content_saving():
    """Test that content is properly saved to the scrapes directory"""
    with tempfile.TemporaryDirectory() as temp_dir:
        output_dir = Path(temp_dir) / "scrapes"
        scraper = ScraperModule(
            delay_between_requests=0.0,  # No delay for testing
            output_dir=str(output_dir)
        )

        # Test with a local HTML content simulation instead of an actual URL
        # Since we can't reliably test with external URLs in unit tests
        # we'll test the file saving functionality separately

        # We'll just test that the output directory is set up properly
        assert output_dir.exists()
        assert output_dir.is_dir()


def test_scraper_types():
    """Test that different scraper types are available"""
    with tempfile.TemporaryDirectory() as temp_dir:
        output_dir = Path(temp_dir) / "scrapes"
        scraper = ScraperModule(output_dir=str(output_dir))

        # Check that the scraper types are registered
        assert 'default' in scraper.scraper_types
        assert 'bs' in scraper.scraper_types      # BeautifulSoup scraper
        assert 'w3m' in scraper.scraper_types    # W3M scraper
        assert 'js' in scraper.scraper_types     # JavaScript scraper


def test_scrape_response_structure():
    """Test the structure of ScrapeResponse"""
    from scraper import ScrapeResponse

    response = ScrapeResponse(
        url="https://example.com",
        status_code=200,
        content="<html><body>test</body></html>",
        headers={"Content-Type": "text/html"},
        response_time=0.1,
        result=ScrapeResult.SUCCESS
    )

    assert response.url == "https://example.com"
    assert response.status_code == 200
    assert response.content == "<html><body>test</body></html>"
    assert response.headers == {"Content-Type": "text/html"}
    assert response.response_time == 0.1
    assert response.result == ScrapeResult.SUCCESS
    assert response.error_message is None
    assert response.filename is None
